moved OrphanedFilesCleaner into DownloadCleaner
