+++ code/backend/Cleanuparr.Infrastructure.Tests/Features/Jobs/DownloadCleanerTests.cs
@@ -52,7 +52,7 @@ public class DownloadCleanerTests : IDisposable
             _fixture.EventPublisher,
             _fixture.TimeProvider,
             _fixture.HardLinkFileService,
-            _fixture.OrphanedFilesCleaner
+            _fixture.DryRunInterceptor
         );
     }
