+++ code/backend/Cleanuparr.Infrastructure.Tests/Features/Jobs/Integration/DownloadCleanerIntegrationTests.cs
@@ -48,7 +48,7 @@ public class DownloadCleanerIntegrationTests : IDisposable
             _fixture.EventPublisher,
             _fixture.TimeProvider,
             _fixture.HardLinkFileService,
-            _fixture.OrphanedFilesCleaner);
+            _fixture.DryRunInterceptor);
     }
 
     /// <summary>
