+++ code/backend/Cleanuparr.Infrastructure.Tests/Features/Jobs/Integration/IntegrationTestFixture.cs
@@ -53,7 +53,6 @@ public class IntegrationTestFixture : IDisposable
     public IDownloadServiceFactory DownloadServiceFactory { get; private set; }
     public IBlocklistProvider BlocklistProvider { get; private set; }
     public IHardLinkFileService HardLinkFileService { get; private set; }
-    public OrphanedFilesCleaner OrphanedFilesCleaner { get; private set; } = null!;
     public INotificationPublisher NotificationPublisher { get; private set; }
     public IDryRunInterceptor DryRunInterceptor { get; private set; }
     public IEventPublisher EventPublisherInterface { get; private set; } = null!;
@@ -135,11 +134,6 @@ public class IntegrationTestFixture : IDisposable
             EventPublisher,
             EventsContext,
             DataContext);
-
-        OrphanedFilesCleaner = new OrphanedFilesCleaner(
-            Substitute.For<ILogger<OrphanedFilesCleaner>>(),
-            DataContext,
-            DryRunInterceptor);
     }
 
     /// <summary>
